Enable IntLongMath checker of error prone
- This checker avoid that an expression of type int may
overflow before being assigned to a long.
- Do not longer use long where it is not needed since an
integer can be used from a dex specification point of view.

Bug:

+++ src/main/java/com/android/tools/r8/ir/optimize/CodeRewriter.java
@@ -494,7 +494,7 @@ public class CodeRewriter {
             }
 
             // Get the existing dex size for the payload and switch.
-            long currentSize = Switch.payloadSize(keys) + Switch.estimatedDexSize();
+            int currentSize = Switch.payloadSize(keys) + Switch.estimatedDexSize();
 
             // Never replace with more than 10 if/switch instructions.
             if (outliers.size() + sequences.size() <= 10) {
